requirements.txt
pandas==2.0.3
pypdf==4.3.1
python_dateutil==2.8.2
Requests==2.32.3
streamlit==1.25.0
trubrics==1.6.2
weaviate-client==4.*
ipykernel
